backend/requirements.txt
numpy
sentence-transformers
django-cors-headers